Move app to /src directory
