+++ src/app/issues/[id]/DeleteIssueButton.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import { Spinner } from "@/app/components"
+import { Spinner } from "@/components"
 import { AlertDialog, Button, Flex } from "@radix-ui/themes"
 import axios from "axios"
 import { useRouter } from "next/navigation"
