+++ src/app/issues/[id]/IssueDetails.tsx
@@ -1,4 +1,4 @@
-import { IssueStatusBadge } from '@/app/components'
+import { IssueStatusBadge } from '@/components'
 import { Issue } from '@prisma/client'
 import { Card, Flex, Heading, Text } from '@radix-ui/themes'
 import ReactMarkdown from 'react-markdown'
