+++ src/app/issues/[id]/page.tsx
@@ -1,4 +1,4 @@
-import authOptions from '@/app/auth/authOptions';
+import authOptions from '@/auth/authOptions';
 import prisma from '@/prisma/client';
 import { Box, Flex, Grid } from '@radix-ui/themes';
 import { getServerSession } from 'next-auth';
