+++ src/app/issues/list/IssueTable.tsx
@@ -1,5 +1,5 @@
 
-import { IssueStatusBadge } from '@/app/components';
+import { IssueStatusBadge } from '@/components';
 import { Issue } from '@prisma/client';
 import { ArrowUpIcon } from '@radix-ui/react-icons';
 import { Table } from '@radix-ui/themes';
