+++ src/app/layout.tsx
@@ -1,10 +1,10 @@
+import AuthProvider from '@/auth/Provider';
+import QueryClientProvider from '@/libs/reactQuery/QueryClientProvider';
 import { Container, Theme } from '@radix-ui/themes';
 import '@radix-ui/themes/styles.css';
 import type { Metadata } from 'next';
 import { Inter } from 'next/font/google';
 import NavBar from './NavBar';
-import QueryClientProvider from './QueryClientProvider';
-import AuthProvider from './auth/Provider';
 import './globals.css';
 import './theme-config.css';
