+++ src/components/Skeleton.tsx
